requirements.txt
duckdb>=0.8.0
pandas>=2.0.0
matplotlib>=3.7.0
tabulate>=0.9.0
python-dotenv>=1.0.0
fpdf2>=2.7.4
Pillow>=9.5.0